Add playRound and playGame functions
I added the two missing functions. The game now
plays fine as long as the unser input is correct.
I still need to figure out how to properly exit
the game when the user cancels out via the prompt
window and how to properly re-prompt for a valid
answer in case of invalid prompt entries.