Complete the game and debug errors
The user interaction is now working as intended.
I always received a value of "undefined" in the
console whenever the user enters an invalid answer.
Reason was that I forgot to put "return" before
re-calling the humanChoice function. I also managed
to get the desired response in case the user
cancels the game by identifying this special case
with the cancel buttons own null value instead of
returning false.